Better mismatch handling for partial paths
In the case of partial matches (where only part of the path is found)
jsonsl doesn't seem to be very adept at filtering out the bad results.
As a result, we add some logic in the POP callback that, when a partial
match is found, ensures that the innermost parent is a suitable type for
the first unfound level.