MB-23522: Add support for setting Result match Loc
With the changes being made to memcached for MB-23522 we need a
facility to modify the result from non-subjson operations. This commit
allows outside parties to set the match location and m_newdoc of a
Result so that the existing subdoc infrastructure within memcached can
continue to be used.

As part of the change, m_newdoc has been changed to a std::array to
provide a bit more safety.

Change-Id: Iea340a24bd924637b1f51644e95d3ab42dd90249
Reviewed-on: http://review.couchbase.org/75862
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>